Setup Tool versioning
The current versions are set as `Tool.Default.*` but now one can add more; and
make then more “typed” (like the `` `Bwa _`` one).

This is the first step of #34 then a tool's version should be part of its
parameters (hence part of the “uniquification” of the file names and exported in
the JSON blob).