Implement “requirements” for `Machine.run_program`
The idea is that users of the `Make_fun.t` functions express their needs with a
`~requirements` argument.

Both the `Make_fun` and `Machine` modules provide functions with default
values.

The `Make_fun.Requirement.t` variants are of course just a first stab to be
refined later (and some cases are not used anywhere).

This “fixes” #163.

We can start taking notes when we run tools, so that they can give a better
estimate of their memory requirements (cf. #166).

I came across #179 but did not fix it yet (I actually don't really know these
tools).