add support for pyensembl
this will allow us to cache genome data sets that are important
for tools like topiary and isovar to be cached up front. For a better
abstraction, the reference genomes now have species and ensembl release
attributes to them. This will force us to define the genome context
explicitly and isolate the caching step into a separate cluster job
when necessary.