Make `EDSL.save`'s behavior explicit in its type
The implementation of `To_workflow.save` does not have the same
semantics as the one in `src/examples/edsl_extension_register_result.ml`
or in Epidisco (cf. [Episave][]).

This commit gives to the result of the function the type
``[ `Saved of 'a ] repr`` forcing users to handle the case explicitly.

[Episave]: https://github.com/hammerlab/epidisco/blob/3b8f722b7913e3835242789eea25258edd6853ac/src/lib/save_result.ml